Division hashing & stepping collision
I used a division hashing technique w/ a stepping collision method. I chose this because It seemed like an easy method to start out with.
The statistics are:
Movie had 260254 collisions, 0 unused spaces and a construction time of .14 seconds.
Quote had 577052 collisions, 0 unused spaces and a construction time of .36 seconds.