Division hashing & list appending collision
I used a division hashing technique w/ a list appending collision method. I chose this because I expected it to reduce the amount of collisions.
The statistics are:
Movie had 14003 collisions, 0 unused spaces and a construction time of .14 seconds.
Quote had 14003 collisions, 0 unused spaces and a construction time of .21 seconds.