Mid-square hashing & list appending collision
I used a mid-square hashing technique w/ a list appending collision method. I chose this method because I expected it to have more variety to the keys, this would hopefully reduce the collisions.
The statistics are:
Movie had 14046 collisions, 43 unused spaces and a construction time of .17 seconds.
Quote had 14080 collisions, 77 unused spaces and a construction time of .21 seconds.