Mid-square hashing & linear probe by prime number collision
I used a Mid-square hashing technique w/ a linear probing collision method. It also was a dynamic hash table, so it expanded as needed. I chose this method because I hoped it would limit the amount of collisions.
The statistics are:
Movie had 5349601 collisions, 2345 unused spaces, a construction time of 5.61 seconds, and the length of the table was 17345.
Quote had 4219907 collisions, 2440 unused spaces, a construction  time of 5.47 seconds, and a length of the table was 17440.